+++ tests/testthat/test-cast.R
@@ -30,3 +30,34 @@ test_that("fallback casts work", {
   character2 <- new_class("character2", "character")
   expect_equal(cast(character2("x"), character), "x")
 })
+
+describe("cast_next()", {
+  it("overrides dispatch, matching exact class", {
+    foo1 <- new_class("foo1")
+    foo2 <- new_class("foo2", foo1)
+    foo3 <- new_class("foo3", foo2)
+
+    bar <- new_generic("bar", "x")
+    method(bar, foo1) <- function(x) 1
+    method(bar, foo3) <- function(x) 3
+
+    expect_error(bar(cast_next(foo3(), foo2)), "Can't find method")
+    expect_equal(bar(cast_next(foo3(), foo1)), 1)
+  })
+
+  it("only affects one dispatch", {
+    foo1 <- new_class("foo1")
+    foo2 <- new_class("foo2", foo1)
+
+    bar1 <- new_generic("bar1", "x")
+    method(bar1, foo1) <- function(x) 1
+    method(bar1, foo2) <- function(x) 2
+
+    bar2 <- new_generic("bar2", "x")
+    method(bar2, foo1) <- function(x) c(1, bar1(x))
+    method(bar2, foo2) <- function(x) c(2, bar1(x))
+
+    expect_equal(bar2(cast_next(foo2(), foo1)), c(1, 2))
+    expect_equal(bar2(cast(foo2(), foo1)), c(1, 1))
+  })
+})